% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copingmethods-package.R
\docType{package}
\name{copingmethods-package}
\alias{copingmethods}
\alias{copingmethods-package}
\title{copingmethods: Tools to Explore Global Mental Health Trends in Depression and Coping Mechanisms}
\description{
The `copingmethods` package offers tools to explore mental health data, focusing on depression rates across age groups and coping mechanisms used by different countries. This package includes two datasets, `depression_age` (depression rates from 1990 to 2019 by age group) and `percent_copingmethods` (coping methods data for 2020). An interactive Shiny app allows users to analyze these trends interactively.
}
\author{
\strong{Maintainer}: Pooja Rajendran Raju \email{praj0022@student.monash.edu}

}
\keyword{internal}
