% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{depression_age}
\alias{depression_age}
\title{Depression Rates by Age Group for Various Countries, 1990-2019}
\format{
\subsection{'depression_age'}{

A data frame with 6420 rows of 18 columns.
\describe{
\item{Entity}{Name of Country}
\item{Code}{Country Code}
\item{Year}{Year of Data Collection}
\item{Age_5_14}{Percentage of respondents aged 5-14 who reported experiencing depression}
\item{Age_15_19}{Percentage of respondents aged 15-19 who reported experiencing depression}
\item{Age_20_24}{Percentage of respondents aged 20-24 who reported experiencing depression}
\item{Age_25_29}{Percentage of respondents aged 25-29 who reported experiencing depression}
\item{Age_30_34}{Percentage of respondents aged 30-34 who reported experiencing depression}
\item{Age_35_39}{Percentage of respondents aged 35-39 who reported experiencing depression}
\item{Age_40_44}{Percentage of respondents aged 40-44 who reported experiencing depression}
\item{Age_45_49}{Percentage of respondents aged 45-49 who reported experiencing depression}
\item{Age_50_54}{Percentage of respondents aged 50-54 who reported experiencing depression}
\item{Age_55_59}{Percentage of respondents aged 55-59 who reported experiencing depression}
\item{Age_60_64}{Percentage of respondents aged 60-64 who reported experiencing depression}
\item{Age_65_69}{Percentage of respondents aged 65-69 who reported experiencing depression}
\item{Age_70}{Percentage of respondents aged 70 and above who reported experiencing depression}
\item{Standardised}{Standardised percentage of respondents across all age groups who reported experiencing depression}
}
}
}
\source{
\url{https://ourworldindata.org/grapher/depressive-disorders-prevalence-by-age}
}
\usage{
depression_age
}
\description{
This dataset provides annual data on depression rates for different age groups across countries from 1990 to 2019. It allows analysis of trends in age-specific depression rates over time.
}
\keyword{datasets}
