% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_app.R
\name{launch_copingmethods}
\alias{launch_copingmethods}
\title{Launch Coping Methods Shiny App}
\usage{
launch_copingmethods()
}
\value{
No return value. This function is called for its side effects (launching the Shiny app).
}
\description{
This function launches the Shiny app named copingmethods_app included in the copingmethods package.
}
